+++ NEWS.md
@@ -11,6 +11,7 @@ Enhancements:
  * Tested with node-0.10 (#141)
  * Use single redis pooler for torque and grainstore
  * Reduce cost of garbage collection for localized resources
+ * Allow limiting number of templates for each user (#136)
 
 1.7.1 -- 2014-02-11
 -------------------